[clean] Refactor
